Changed GenerateOtp from int to long
Since some token use as seed values that are outside the range of int,
changed the signature of the GenerateOtp method in
CounterBasedOtpGenerator to long from int

+++ OATH.Net/CounterBasedOtpGenerator.cs
@@ -75,7 +75,7 @@ namespace OathNet
         /// </summary>
         /// <param name="counter">The counter value to use.</param>
         /// <returns>The OTP for the given counter value.</returns>
-        public virtual string GenerateOtp(int counter)
+        public virtual string GenerateOtp(long counter)
         {
             var text = BitConverter.GetBytes(counter);
